Separate common parts of FFI for WASM use
